My introduction
Hi. I have been interested in 3D printing for some time and at 65 years old, some of my friends and such are wondering what I'm thinking. I got a printer and have worked with it ok using other peoples files but I want to make my own. Thats the whole idea of this was to make stuff I want to make to share with other like minded folks. I'm not going to try and make a living at this. I need things to do now that I have been put out to pasture so to speak and in my own space. I'm having trouble understanding software and tutorials that have differences that I'm not tech savy enough to understand and know how to do, so I am frustrated and confused about some of it. This meshmixer sounds like it is something I should try to work with so I bought it. Now we'll see what I can do with it. Love to chat with some new folks about their tips and help. Sorry I have nothing much to offer back unless you live close enough to get together and work on this stuff.